requirements.txt
python-telegram-bot==9.0.0
beautifulsoup4==4.6.0
get==0.0.39
lxml==4.1.1
numpy==1.14.0
post==0.0.26
public==0.0.65
query-string==0.0.28
urllib3==1.22
